api/af-south-1/requirements.txt
cffi==2.0.0
cryptography==46.0.5
pycparser==3.0
pyjwt==2.11.0
